Move to contracts crate for easier sharing
